feat: high order differentiation
