+++ DeZero/__init__.py
@@ -4,6 +4,6 @@ from DeZero.core import using_config
 from DeZero.core import no_grad
 from DeZero.core import as_array
 from DeZero.core import as_variable
-from DeZero.functions import setup_variable
+from DeZero.core import setup_variable
 
 setup_variable()
