+++ DeZero/functions.py
@@ -36,109 +36,6 @@ def square(x):
     return f(x)
 
 
-class Add(Function):
-    def forward(self, x0, x1):
-        y = x0 + x1
-        return y
-
-    def backward(self, gy):
-        return gy, gy
-
-
-def add(x0, x1):
-    x1 = as_array(x1)
-    return Add()(x0, x1)
-
-
-class Mul(Function):
-    def forward(self, x0, x1):
-        y = x0 * x1
-        return y
-
-    def backward(self, gy):
-        x0, x1 = self.inputs[0].data, self.inputs[1].data
-        return gy * x1, gy * x0
-
-
-def mul(x0, x1):
-    x1 = as_array(x1)
-    return Mul()(x0, x1)
-
-
-class Neg(Function):
-    def forward(self, x):
-        return -x
-
-    def backward(self, gy):
-        return -gy
-
-
-def neg(x):
-    return Neg()(x)
-
-
-class Sub(Function):
-    def forward(self, x0, x1):
-        y = x0 - x1
-        return y
-
-    def backward(self, gy):
-        return gy, -gy
-
-
-def sub(x0, x1):
-    x1 = as_array(x1)
-    return Sub()(x0, x1)
-
-
-def rsub(x0, x1):
-    x1 = as_array(x1)
-    return Sub()(x1, x0)
-
-
-class Div(Function):
-    def forward(self, x0, x1):
-        y = x0 / x1
-        return y
-
-    def backward(self, gy):
-        x0, x1 = self.inputs[0].data, self.inputs[1].data
-        gx0 = gy / x1
-        gx1 = gy * (-x0 / x1 ** 2)
-        return gx0, gx1
-
-
-def div(x0, x1):
-    x1 = as_array(x1)
-    return Div()(x0, x1)
-
-
-def rdiv(x0, x1):
-    x1 = as_array(x1)
-    return Div()(x1, x0)
-
-
-# 累乗の計算
-# 指数は定数として扱い、勾配は追跡しない
-class Pow(Function):
-    def __init__(self, c):
-        self.c = c
-
-    def forward(self, x):
-        y = x ** self.c
-        return y
-
-    def backward(self, gy):
-        x = self.inputs[0].data
-        c = self.c
-        gx = c * x ** (c - 1) * gy
-        return gx
-
-
-def pow(x, c):
-    return Pow(c)(x)
-
-
 class Sin(Function):
     def forward(self, x):
         y = np.sin(x)
@@ -165,15 +62,3 @@ def my_sin(x, threshold=0.0001):
             break
     return y
 
-
-def setup_variable():
-    Variable.__add__ = add
-    Variable.__radd__ = add
-    Variable.__mul__ = mul
-    Variable.__rmul__ = mul
-    Variable.__neg__ = neg
-    Variable.__sub__ = sub
-    Variable.__rsub__ = rsub
-    Variable.__truediv__ = div
-    Variable.__rtruediv__ = rdiv
-    Variable.__pow__ = pow
